Set the render start time for all pending event entries
We are working on recording the INP breakdown metrics, and we want to
record the rendering start delay and rendering duration. We need the
timestamp of render start to calculate these time durations. In this CL, we are adding the render start time into all the pending event entries'
EventTimingReportingInfo.

Bug: 40858679
Change-Id: I1930e9720e78ee25c0627f0b992f817a97809fd0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6055328
Reviewed-by: Michal Mocny <mmocny@chromium.org>
Commit-Queue: Lan Wei <lanwei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1391840}
NOKEYCHECK=True
GitOrigin-RevId: 2b08e941f979b78e54ed0e98538c7e2e80803a0b