Ship the X25519 WebCrypto algorithm
The intent-to-ship request [1] has been approved, so this CL we are going to enable by default the WebCryptoCurve25519 runtime flag.

Additionally, this CL also splits up the WebCryptoCurve25519 flag, assuming it applies only to the X25519 feature, and defines a new one for the Ed25519 algorithm. It was initially planned to ship both together, but finally it was not ready for the intent.

[1] https://groups.google.com/a/chromium.org/g/blink-dev/c/A7lbRONS1lY/m/6J6wRdj4AAAJ

Bug: 378856322
Change-Id: I97c12bbfb843b278ec216070e373f6641a20a7e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6070029
Commit-Queue: Javier Fernandez <jfernandez@igalia.com>
Reviewed-by: Rick Byers <rbyers@chromium.org>
Reviewed-by: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1391915}
NOKEYCHECK=True
GitOrigin-RevId: 3136845f1fb5146ca2b608446bd693712011c827