Fix IDL-exposed const-ness on FileSystemChangeRecord methods.
This stops using const a context where the bindings code doesn't
guarantee that const-ness in any way.

This is one of the C++ fixes needed to turn this sort of incorrect
function signature into a compilation error.

For this error in particular I think the C++ code is substantively
misleading since javascript callers would be able to mutate the returned
objects.

Bug: 40275482
Change-Id: Iad84dd95227b6d95d6120055b2d349147d70d060
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6067868
Commit-Queue: David Baron <dbaron@chromium.org>
Reviewed-by: Nathan Memmott <memmott@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1392214}
NOKEYCHECK=True
GitOrigin-RevId: 2bf0aa292ef4096f5b31653c40dd3b5977017f7c