Add devtools protocol to compute values in the context of the element
Add devtools protocol for computing the specified values in the context
of the provided element. For example, a value of '1em' is evaluated
according to the computed 'font-size' of the element. More info about
API in the doc [0].

This CL does not resolve percantages, it will be done separately after.
Also the protocol does not work for the pseudo elements that does not
produce an element (see test testPseudoElementsNoElement for example),
this part needs further discussion.

[0] https://docs.google.com/document/d/1zyKdPREtKT8OU4WtlHV_Wxet3SvyUtAXrTdFLPmYmdU

Change-Id: Ib2cafbc018f3e474d35632884c79d7ce833e8bef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6012406
Commit-Queue: Munira Tursunova <moonira@google.com>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1392472}
NOKEYCHECK=True
GitOrigin-RevId: c35cd211855fb67e1969da0589e3ddf6024c6321