Make Chrome with B&A k-anonymity support work with older servers
Change Chrome to still accept B&A responses from Bidding and Auction
servers that don't support k-anonymity. This is intended to ease the
launch of the Chrome k-anonymity feature by not requiring all of the
servers to upgrade before the Chrome launch.

This change also fixes a bug where the buyerAndSellerReportingId was
checked for k-anonymity based on the on-device instead of relying on
the k-anonymity check performed on the server.

Bug: 41495800
Change-Id: I27c0ea0f871d86dc87870ecbdad408e79d724ea5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6063083
Reviewed-by: Qingxin Wu <qingxinwu@google.com>
Commit-Queue: Russ Hamilton <behamilton@google.com>
Cr-Commit-Position: refs/heads/main@{#1392565}
NOKEYCHECK=True
GitOrigin-RevId: 1cd4c5e26d8e49dc0151e3887d8dcdbc98c57882