Warn about the expiry date of certificates later
This is a temporary change, until the certificates are regenerated.
The certificates did not expire yet, they expire in ~30 days. This code
gave an error message if the certificates are going to expire in less
than 30 days. As a result, 12 CI bots are consistently failing now.

This CL makes the code give an error message if the certificates are
going to expire in less than 20 days.

This CL makes it possible for the bots to pass, as a temporary solution,
until the certificates are regenerated.

The process of regenerating certificates seems to be very complicated.
This CL is simple and has owners override so that the issue is fixed
fast. I will regenerate the certificates in an immediate follow-up CL,
which I will submit only after owners review it.

Bug: 382650353
Change-Id: I4cd94cd3508e9ceecfd786f5fdc4179d93984b9a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6074479
Commit-Queue: Jan Keitel <jkeitel@google.com>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Auto-Submit: Theo Cristea <theocristea@google.com>
Commit-Queue: Theo Cristea <theocristea@google.com>
Owners-Override: Theo Cristea <theocristea@google.com>
Cr-Commit-Position: refs/heads/main@{#1392851}
NOKEYCHECK=True
GitOrigin-RevId: b2e7f2d261115c2f6c50cff373c3551c31d8cea8