blink::HTMLVideo/CanvasElement: Set image rendering and dynamic range
In HTMLCanvasElement, propagate DynamicRangeLimit along all of the
paths where FilterQuality is propagated.

In HTMLVideoElement, do the same propagation. This matches the behavior
of HTMLCanvasElement with:
* HTMLVideoElement::SetCcLayer matching HTMLCanvasElement::
  InitializeLayerWithCSSProperties
* HTMLVideoElement::StyleDidChange matching HTMLCanvasElement::
  StyleDidChange

In LayoutVideo, override the StyleDidChange (as is done in
LayoutHTMLCanvas), and propagate the old and new styles to
HTMLVideoElement::StyleDidChange (as is done in HTMLCanvasElement).

There is some code replication between these two structures. This could
be unified by tracking the DynamicRangeLimit and FilterQuality in
LayoutImage. The actual gain would be fairly limited (this would not
be the case if LayoutImage had access to the relevant cc::Layer, but
this is not obviously desirable).

This will successfully plumb this information through to the
viz::SkiaRenderer.

Bug: 40843521, 40277822
Change-Id: I4f94c1a906a1c18e144ac3282e005520e5eb6bc3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6062879
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1392983}
NOKEYCHECK=True
GitOrigin-RevId: a366a3a6602da9cce85109c0fed565182ab70bc9