Remove the workaround for crbug.com/362676838
There was an issue that any ongoing sessions will be aborted once the
language detection model is loaded. To avoid the issue,
workaround-for-362676838.js was introduced.

The issue has been fixed. Also workaround-for-362676838.js is creating a
AIWriter which doesn't work with the latest base model.

So this CL removes the workaround.

Bug: 362676838, 378549335
Change-Id: I761c03c3c653ec6c7903601f65f462cf7a53b4c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6074176
Reviewed-by: Mingyu Lei <leimy@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1393466}
NOKEYCHECK=True
GitOrigin-RevId: 6cac3e5a6af780cf1e28db941e99f1312c2f8424