Switch JS LanguageDetection to request a model directly from browser.
Previously it was using the model managed Chrome-UI-translate. This
was a convenient hack but that model is not always present. With this
change, the blink code owns its own model and attempting to doing
detection from JS will cause the model to be populated.

This introduces the blink platform LanguageDetectionModel which knows
how to get a model file from the browser and uses that to perform
language detection.

Bug: 354069716
Change-Id: Id68ac955e652bec8f5c80dd9e7de1842dcfddac3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6048231
Reviewed-by: Tsuyoshi Horo <horo@chromium.org>
Auto-Submit: Fergal Daly <fergal@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1393616}
NOKEYCHECK=True
GitOrigin-RevId: 2db39fe95326b827c758463027db322c02377a1d