[css-nesting] Implement CSSScopeRule::SetPreludeText by rule replacement
CSSScopeRule::SetPreludeText is a function that replaces the prelude
of some @scope rule. It is not web-exposed (i.e. not bound to any JS
API), but it's used by Devtools to edit @scope rules dynamically.
This function is currently implemented as modifying the contents
of the StyleRuleScope in-place, which is an approach we're moving
away from (see other work related to Issue 363019837).

This CL instead implements the mutation as replacing the whole
StyleRuleScope held by CSSScopeRule. This fixes nullptr-crashes
for invalid preludes, but there should otherwise not be any behavior
change.

Bug: 363019837
Change-Id: Iefdc96b334c26723ad4164045f741d64825fde01
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6069931
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1393632}
NOKEYCHECK=True
GitOrigin-RevId: ed46557e198a8fce6c1ef52b38e5c17734c99ba9