Pass parameters directly to DrawTextInternal()
Currently, DrawTextInternal() can uses different values for TextAlign,
TextBaseline and Font depending on whether the optional TextCluster
parameter is passed. Additionally, it internally sets values for the
start and end index to render from the input string, which match the
whole string unless a cluster is passed.

With this CL, DrawTextInternal() now expects the font, align, baseline,
begin index, and end index to be passed as parameters. This simplifies
its internal logic and moves that responsibility to the caller; and
removes the need for the TextCluster parameter.

Font is passed a pointer that is null when there is no TextCluster,
since the timing of the call to AccessFont() is important for updating
identifiability_study_helper_. If AccessFont() is called before
DrawTextInternal(), a unit test fails as a result
(https://crsrc.org/c/third_party/blink/renderer/modules/canvas/canvas2d/canvas_rendering_context_2d_api_test.cc;l=838)

In the next CL in the stack, this is used to facilitate the use of an
option parameter for ctx.fillTextCluster() that can override some of
this values.

Bug: 341213359
Change-Id: I0115a7c6ca991b5b0a1e07cac720976af8b27ca0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6035867
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Commit-Queue: Andres Ricardo Perez <andresrperez@chromium.org>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1393897}
NOKEYCHECK=True
GitOrigin-RevId: e2439f4a12972b455e3c9e0618f83fe6c2848888