Add use counter for touch action handwriting keyword
This CL adds a Blink use counter that gets fired when a web page has an
element whose computed style for the touch-action css attribute would
disable handwriting for the element if the handwriting keyword were
implemented.

Handwriting would be disabled when touch-action is specified with any
keyword other than `auto` or `manipulation`. The use counter will be
fired when any other keyword[1] is specified, excepting when
touch-action is declared with the `none` keyword in which case
handwriting would be disabled but that would be by design.

For more context on the cause of this use counter, please see the linked
bug.

[1] https://w3c.github.io/pointerevents/#dfn-touch-action-values

Bug: 378027646
Change-Id: I88e9c1641490382cc8477cb59e97ae671c680ea9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6005047
Reviewed-by: Adam Ettenberger <Adam.Ettenberger@microsoft.com>
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Gaston Rodriguez <gastonr@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1393914}
NOKEYCHECK=True
GitOrigin-RevId: f5c74d45267a7d9eaae001041b672120f9634d70