webgpu: Enable compatibility mode with featureLevel
This change ensures that featureLevel correctly enables compatibility
mode when it's enabled by flags.  It also adds a warning for developers
when featureLevel is used to request compatibility mode but is ignored
because compatibility mode is not enabled by flags. This will help
ensure a smooth transition when featureLevel is fully enabled.

Bug: 366151404
Change-Id: I894f09006725d73ef643acfd5d0131bda81f5aa9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6075283
Commit-Queue: Fr <beaufort.francois@gmail.com>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Yoav Weiss (@Shopify) <yoavweiss@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394170}
NOKEYCHECK=True
GitOrigin-RevId: a13c6397ae20180cef2d0d25327dfe2289334d36