Skip ruleset invalidation of non-rendered shadow tree
If the host is not rendered, invalidation of the shadow tree will not
have any effects. Invalidation of the host causing it to be rendered
will make sure any descendants gets style as needed too.

This fixes one of the instances where
StyleRecalcContext::FromInclusiveAncestors is invoked for an element
with a null ComputedStyle.

Bug: 40250356
Change-Id: I7b2a8e4319204d7b818084e54504141709f7ee12
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6082253
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394332}
NOKEYCHECK=True
GitOrigin-RevId: 7f04348dc96b5a412ffdadda983a8e68c8600f94