Use private_aggregation_promise_test() in fenced-frame WPTs
It's a good idea to use `private_aggregation_promise_test()` because it
automatically resets the reporting endpoints before running the provided
test callback.

It's entirely possible that this change is a no-op. Polling for reports
is a destructive operation on the server, so if every test polled for
reports at the end, then explicitly clearing the stash would never be
necessary.

Bug: 362901607
Change-Id: I9ddc44e2a8cd4534ed15c8b9d3295bc04721cac0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6085835
Commit-Queue: Dan McArdle <dmcardle@chromium.org>
Reviewed-by: Alex Turner <alexmt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394339}
NOKEYCHECK=True
GitOrigin-RevId: ee7cc24ceba333b586fc88dbc64b194cdc1d8e18