Include call time of performance.measure and mark in trace event
The timestamp of the event it self is overridden to be the start time
passed as argument to the API, meaning the time in which the API was
called is lost. This CL adds the call time as part of the payload of the
trace events dispatched for performance.measure and mark.

This will be useful in improving the stack traces we show in the
Performance panel in DevTools for user timings, as we will be
consolidating the stack data we collect from tracing and samples, for which a event timestamp will be necessary.

note: I chose "now" as the carrying property to distinguish it from the
"startTime" property of performance marks / measures, since I thought
"timeStamp" doesn't make the difference clear enough.

Bug: 381861194
Change-Id: I8ad5262e80596b324eee2b1df0dc558f9de13f19
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6074543
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Annie Sullivan <sullivan@chromium.org>
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Yang Guo <yangguo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394716}
NOKEYCHECK=True
GitOrigin-RevId: 754400c091886bb6b7234f6820bf661c2c361a62