Add a constexpr ExceptionContext constructor for the modal case.
For the mos commont case by far (where we use a default empty
context that shouldn't ever be read), use a constexpr.

The other cases are promise-returning APIs that set an
API-specific context. Change the constuctor to expect the
caller to create the ExceptionContext, instead of taking the
parameters needed to construct the ExceptionContext.

Change-Id: I7c0c95a13267da8e28dea31de8bb272a7d464a4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6078343
Commit-Queue: Nate Chapin <japhet@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394928}
NOKEYCHECK=True
GitOrigin-RevId: 6a0b666a1f2db769934788ec2d571aee368df0eb