Prune superfluous calls to SetCompositorPending
We only need to call SetCompositorPending with the pending cancel
reason if the animation is running on the compositor.

The stack trace on the bug report showed a significant time being
spent in HasProperty. The timing was also optimized in this CL to
prevent unnecessary duplicate calculations.

Bug: 382394791
Change-Id: I03ffa1b486b267e05f63328212d192dfca26eb53
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6076354
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Reviewed-by: Claire Chambers <clchambers@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1395390}
NOKEYCHECK=True
GitOrigin-RevId: 1282289030ab2026a8db4fb82dc6e20d3bf028be