Remove third_party/pexpect as unused and unowned
We need OWNERS for all of our third_party dependencies, and the
OWNERS of pexpect are no longer active in Chromium. It was originally added in 2012 (crbug.com/40951150) to make some performance infrastructure work on MacOS but hasn't really been touched since and appears not to be needed anymore.

If I missed something and this needs to come back, it's fine to
revert this. We just need to find 2+ OWNERS to replace those listed
here.

Also removes lighthttpd_server.py since it relies on pexpect but is used only by legacy cronet tests which are not currently being used.

Bug: 378273938
Change-Id: Ib1ae6eb06ad5578826f9e741c29a7a2c4383c6e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6077590
Reviewed-by: Mohannad Farrag <aymanm@google.com>
Commit-Queue: Rick Byers <rbyers@chromium.org>
Reviewed-by: Sam Maier <smaier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1395433}
NOKEYCHECK=True
GitOrigin-RevId: 01656c8d36fe531399d2b235747cb31961f36497