iwa: Fix crash when permissions policy header vs manifest value differ
This fixes a crash in IWAs when a Permissions-Policy header references
a feature not declared in the IWA's manifest.

This also adds unit test coverage to the
PermissionsPolicy::CombinePolicies function and fixes a bug the tests
uncovered with how 'self' was handled.

Bug: 40236903
Change-Id: I78d7da70155edab38328f810016986441558ac6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6082223
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1395494}
NOKEYCHECK=True
GitOrigin-RevId: c3f72230abeae66c07e19a7d0081c09bc5fcd49d