[Autofill] Fix ListedElements() after insertion of a shadow DOM tree
Autofill uses a variant of HTMLFormElement::ListedElements()
to extract shadow-including listed elements of a form.

This CL adds a missing cache invalidation when an entire
subtree that includes a shadow DOM is inserted into a form.

For example, when JS dynamically turns
  <form>
into
  <form>
    <div>
      #shadow-root
        <input>
then the <form>'s cache needs to be invalidated to that
ListedElements(/*include_shadow_trees=*/true)
includes the <input>.

Bug: 347059988
Change-Id: Ifc2d8e7ea29f9538a7a7c8083883463834cc8677
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6088336
Reviewed-by: Mason Freed <masonf@chromium.org>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Commit-Queue: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/main@{#1395612}
NOKEYCHECK=True
GitOrigin-RevId: 51e2da7845b1c03198a799bc153010e11f17650a