[WebDX] Map web feature WebAppWindowControlsOverlay to webdx feature
The current WindowControlsOverlay UseCounter is fired when the manifest
is parsed so it does not necessarily indicate that the feature is used.
However, WebDXFeature UseCounter is intended to be fired when a web
developer uses the feature so the current use counter can be mapped.

See doc:
https://docs.google.com/document/d/1iiat-77KuFxEWLqSXvC754LG36H_vcbF5A0Y5bwQzVg/edit?tab=t.0

Bug: 366475997
Change-Id: I2d7d546bd4b01eacb9b1f075116c3a1492f6a27e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6089556
Reviewed-by: Lu Huang <luhua@microsoft.com>
Commit-Queue: Hassan Talat <hatalat@microsoft.com>
Reviewed-by: Sam Fortiner <samfort@microsoft.com>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1395752}
NOKEYCHECK=True
GitOrigin-RevId: c3365d25d2f0a4a1fa6a4823887567ff9c9270e3