Add ANCHOR attributes to AnnotatedPageContent.
Links are represented by an AIPageContentAttributeType of kAnchor and
include a URL and rel attribute.

Originally we were thinking of adding url to text_info, and not making
links their own ContentNode. But consider the following example:
<a href="https://www.google.com"><b>G</b>oogle</a>

The link node contains multiple LayoutText objects inside of it. It is
more condensed to have the url be set as a ContentAttribute on its
own ContentNode rather than attaching it to each text info.

Bug: 382119737
Change-Id: I73cfcc5f28e7871b2ea273b1b2c8a2912d41d19f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6072788
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Abigail Klein <abigailbklein@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Michael Crouse <mcrouse@chromium.org>
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1395933}
NOKEYCHECK=True
GitOrigin-RevId: db254d750d5f3cc1e38b0be644e66b2828783a79