[a11y] Align AxBlockFlowIterator behavior with AITB for ruby content
This change adjusts the AxBlockFlowIterator
algorithm to mirror the behavior of the
AbstractInlineTextBox algorithm when handling ruby
content.

The AITB algorithm has an asymmetry that occurs
with ruby-like content (or display: ruby), where
the fragmented structure can lead to scenarios
where an element can compute its "next" sibling
but not its "previous". This is because the
InlineCursor restricts navigation to the same
line, and ruby content can have lines nested
within other lines. See linked bug for an example
and a deeper explanation.

This commit does not yet fix the underlying
asymmetry issue, but it lays the groundwork for a
consistent and comparable behavior between
AxBlockFlowIterator and AbstractInlineTextBox.

Bug: 383384904
Change-Id: I92239d5a2759b27f8201240e1623556d07c41bbf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6089470
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1396112}
NOKEYCHECK=True
GitOrigin-RevId: 7a5ffcced608512ab7eacbf9639bf927c2f651ab