Skip ResourceLoadObserver callbacks when DevTools not open
This CL follows the approach in https://crrev.com/c/5756997 and applies
the change to all requests, not just transparent placeholder images.
The rationale for this is that if the DevTools Network panel is not
open, it is not interested in these requests, and hence we can skip
these callbacks to maximize performance.

Bug: 348442535
Change-Id: I2c3cb55a08d31827935f7f7b04be8c74ce836b31
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6069287
Auto-Submit: Nidhi Jaju <nidhijaju@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Commit-Queue: Takashi Toyoshima <toyoshim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1396546}
NOKEYCHECK=True
GitOrigin-RevId: a520b6b12ea401c8b4f212f504b40b2b44b86816