Add a metric for FIFO underrun in AudioDestination
This CL adds a new metric for FIFO underrun from the dual-threaded
rendering setup in AudioDestination. It records the metric with an
interval of 1000 callbacks. If an audio stream is a short-lived
(less than 1000 callbacks), the metric will be reported at the
destruction of the reporter object.

The aggregate callback implementation (and other components) is
largely based on https://crrev.com/c/4194273, and this is to make
the metric comparison easier across the audio stack.

Bug: 376675150
Change-Id: I99610c8eb4d16236b0a0b048a54f2d2d902841a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6076494
Reviewed-by: Fredrik Hernqvist <fhernqvist@google.com>
Reviewed-by: Michael Wilson <mjwilson@chromium.org>
Commit-Queue: Hongchan Choi <hongchan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1396837}
NOKEYCHECK=True
GitOrigin-RevId: ed6dab7a0b75333b3fe6701ad17dbdab33035394