Add the trigger point of eWallet payment flow
The eWallet payment flow was only triggered when a "payment" rel is
parsed as current implementation. In this case, the trigger point could
happen before href attribute is created and the link element is attached
to the document.

This CL add two more trigger point:
1. When parsing href attribute
2. When the link element is attached to the document.

And the eWallet flow will only be triggered when:
1. rel is "payment".
2. href is not empty.
3. the link element is attached to the document.

We don't need to worry about that the payment flow will be triggered
multiple times, since we have the restriction in `document.cc` to
triggered the flow only once per document.

Bug: 383569850
Change-Id: I3d32e9702fd8c2119af66a90ae83217911f7d619
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6087989
Commit-Queue: Junhui He <junhuihe@google.com>
Reviewed-by: Mason Freed <masonf@chromium.org>
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1396838}
NOKEYCHECK=True
GitOrigin-RevId: 2d0391da75ca4e865dcc1688c6781fb6daba10b4