[Grid] Return empty string for certain grid-template shorthands
In https://chromium-review.googlesource.com/c/chromium/src/+/6068810,
it was pointed out after the change landed that the behavior
implemented there is incorrect. In particular, we should be returning
the empty string when there is a mismatch in row definitions between
grid-template-rows and grid-areas.

This CL updates our behavior to pass the updated test and simplifies
the logic slightly.

Bug: 380369291, 383369462
Change-Id: Idea0974ea1fccbc2ada9bf6a285d6b98c62bb6ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6092205
Commit-Queue: Kurt Catti-Schmidt <kschmi@microsoft.com>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Reviewed-by: Oriol Brufau <obrufau@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1396917}
NOKEYCHECK=True
GitOrigin-RevId: 79593b7ba74d55f5d01e3434fed257e61a676f38