[WebAudio] MediaPlaybackWhileNotVisible Permission Policy Integration
This CL implements a prototype to integrate WebAudio with the
MediaPlaybackWhileNotVisible Permission Policy:
https://github.com/MicrosoftEdge/MSEdgeExplainers/blob/main/IframeMediaPause/iframe_media_pausing.md

The implemented behavior for an iframe with an AudioContext the
permission policy is disabled is:
1. If the AudioContext is running and the iframe is not visible
anymore, the AudioContext will transition to the "interrupted" state.

2. If the iframe is hidden, the AudioContext will not be allowed to
start and remain in the "suspended" state. Attempting to call resume()
on this AudioContext will put it into the "interrupted" state.

3. If the iframe is not hidden and the AudioContext is suspended,
changes in visibility should not yield any state changes.

This CL also contains web tests to check that the behavior above happens
as expected

Bug: 351354996, 374805121
Change-Id: I401db88d3e001de2b0983c532adb6f1db5f29105
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6074328
Reviewed-by: Michael Wilson <mjwilson@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Gabriel Brito <gabrielbrito@microsoft.com>
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1397098}
NOKEYCHECK=True
GitOrigin-RevId: a1c9218b7404043e334327f5469a6dd290a37585