[Editing] Prevent crash when no <table> tag is present during selection
If <td> or some other tags are added using script, then it won't be
converted into a table. This results in crash as |last_closed| will
be assigned as null. Hence, added an if condition to check whether
table part element is inside table element or not in
|style_markup_serializer.cc|.

Bug: 382120192
Change-Id: Ia58dbd543a26033c8f7cf1dc8f32e532f32431d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6090788
Reviewed-by: Siye Liu <siliu@microsoft.com>
Commit-Queue: Shweta Bindal <shwetabindal@microsoft.com>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1397150}
NOKEYCHECK=True
GitOrigin-RevId: 96ebe04158127f1370a0800d2eccad56be513c89