[css-properties-values-api] Test unit arithmetic separately
A subtest relying on unit arithmetic snuck into one of the WPTs
covered by the "Custom Properties" focus area of Interop 2024.
As unit arithmetic is not covered by Interop 2024, the default
resolution to this problem is to remove the whole file from Interop [1].

However, in this case, it's relatively easy and harmless to move
the offending test to a separate file.

[1] https://github.com/web-platform-tests/interop/issues/604

Change-Id: Ic444d07288c2c8bb54595c61813d0217b3cc1032
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6097816
Reviewed-by: Daniil Sakhapov <sakhapov@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1397239}
NOKEYCHECK=True
GitOrigin-RevId: c7fd27c0f28e97ec37eb2c2aecdcd673565075b3