[css] Small optimization in AnonymousNameSetter
The code first checked if a v8::Value was a v8::Number, but then did not
use this knowledge to get the value of the v8::Number. With this CL the
value gets extracted directly, so that no ExceptionState is needed
anymore.

BUG=377724743

Change-Id: Ic53cf41c15cf1236b5c496345a69162ed741a694
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6097456
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1397269}
NOKEYCHECK=True
GitOrigin-RevId: 3b11f57b74b9edab22317e1e803c08691b0420a4