Warn for unsupported gradient color spaces
These four spaces are in the spec, but are unsupported by Skia for
interpolation. We either need to implement them in Skia or inject a
shader. Before this CL the code reaches a NOTREACHED in
Blink::Gradient::ResolveSkInterpolation() and crashes (see linked bug):

https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/platform/graphics/gradient.cc;l=243?q=gradient.cc&sq=package:chromium&ct=os

For the time being, warn the dev and fail as if it's a parsing failure.

Bug: 382086298
Change-Id: Ie47a1caee682e021fdf5df331c50a6e49fcfa050
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6085127
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Peter Boström <pbos@chromium.org>
Commit-Queue: Aaron Krajeski <aaronhk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1397419}
NOKEYCHECK=True
GitOrigin-RevId: f5074d96d7f8930f9a17ff3299682e80a54cdacd