Implement input:open
The :open pseudo-class is specced to work on input elements which
support pickers, so this patch implements :open for input elements.

I manually tested that this works on android which has different code
paths for pickers. I didn't find any tests for these pickers on android
in the first place, except for <input type=file>, based on this:
https://chromium-review.googlesource.com/c/chromium/src/+/6075965

Bug: 324293874
Change-Id: I8a638d0a3ce7edaf39bd2bc9655c0b3aba0f62fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6051109
Reviewed-by: David Baron <dbaron@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1397435}
NOKEYCHECK=True
GitOrigin-RevId: 833248d9d46ef60ec15ec0f1bfe2dcdac13305d0