CanvasColorParams: Clean up helper functions
Add comments at the beginning of CanvasColorParams indicating that it
is used for storing state common to CanvasRenderingContext2D and
OffscreenCanvasRenderingContext2D.

Note that it does not belong in the superclass BaseRenderingContext2D
because it does not do not apply to the PaintRenderingContext2D
subclass (because it does not have a backing bitmap).

Remove the instance of CanvasColorParams in
CanvasRenderingContext, because it was never used (and was hidden
by members of two sub-classes, CanvasRenderingContext2D and
OffscreenCanvasRenderingContext2D).

Remove support for OpacityMode in CanvasColorParams because it isn't
used (it may have been used when CanvasColorParams was doing many other
things that it doesn't do anymore).

Remove CanvasColorParams methods that return strings for enums. Use
the helper function PredefinedColorSpaceToV8 that uses the V8 enums,
and add a static helper CanvasPixelFormatToV8 for the one place it is
needed.

Remove additional unused helper functions in CanvasColorParams.

Bug: 40277213
Change-Id: Ia656313f14567f1819c389bb510ae3977daf0d85
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6105439
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1398268}
NOKEYCHECK=True
GitOrigin-RevId: c1f91d38043e83f44d9642c601cfb40f7efe99a3