Avoid "mutable constants" if char* overload for histograms is removed.
Because "const char*" is not a compile-time constant, it will be
materialized if it's ODR-used. In trying to remove the char* overloads
for histogram functions, I did that and got hit with a trybot failure.

This is solvable by making these `constexpr const char*`, but I then
did the following:
* Inlined them into the call sites
* Changed from arrays to spans for safety
* Since I then had to invoke the explicit span constructor, went ahead
  and changed to string_views to invoke the desired histogram functions
  directly

Bug: 364987728
Change-Id: I1df4fbc03886b3622cd356b7fc7d9cb9cb56d53a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6095367
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1398276}
NOKEYCHECK=True
GitOrigin-RevId: 8aafc68b47b9bb5abc8dd2303460f50a3ecf4fe3