[WebAudio] Don't pull more data than is in the FIFO when in bypass mode
The call to Render() may be interrupted if the device is stopping.  In
this case, we can report an underrun when we try to pull the requested
amount of data.  But it's not really a performance-related underrun,
so instead we should only return the data that is available.

Bug: 376672971
Change-Id: If45fe9f3727bf4aac4a52f9ed8f9b5beba587b24
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6089758
Commit-Queue: Michael Wilson <mjwilson@chromium.org>
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1398285}
NOKEYCHECK=True
GitOrigin-RevId: f971d48706707dac37df80d2469c6110af7bb937